Подключение к БД в основном файле
